refactor: renamed ApplicationsToApprove.tsx -> ApplicationsList.tsx
