+++ src/pages/[domain]/admin/applications/index.tsx
@@ -1,5 +1,5 @@
 import { RoundAdminLayout } from "~/features/admin/layouts/AdminLayout";
-import { ApplicationsToApprove } from "~/features/applications/components/ApplicationsToApprove";
+import { ApplicationsToApprove } from "~/features/applications/components/ApplicationsList";
 
 export default function ApplicationsPage() {
   return (
